+++ R/reconstruct_formulas.R
@@ -1,7 +1,10 @@
 
 
 #' @title Reconstruct model formulas from results tibbles
-#' @description In the results tibble from cross_validate and validate,
+#' @description
+#'  \Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}
+#'
+#'  In the results tibble from cross_validate and validate,
 #'  the model formulas have been split into the columns Dependent, Fixed and Random.
 #'  Quickly reconstruct the model formulas from these columns.
 #' @return