+++ R/validate.R
@@ -2,8 +2,11 @@
 if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
 
 #' @title Validate regression models on a test set
-#' @description Train gaussian or binomial models on a
-#'  full training set and validate it by predicting the test/validation set.
+#' @description
+#'  \Sexpr[results=rd, stage=render]{lifecycle::badge("stable")}
+#'
+#'  Train gaussian or binomial models on a full training set and validate it by
+#'  predicting the test/validation set.
 #'  Returns results in a tibble for easy reporting, along with the trained models.
 #' @inheritParams cross_validate
 #' @param train_data Data Frame.